refactor: move floor selector and control components outside ui
